updated some
